requirements.txt
Flask==3.1.0
Flask_Bcrypt==1.0.1
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
